pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "acres==0.5.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "ci-info==0.3.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "colorclass==2.2.2",
    "comm==0.2.3",
    "compressed-rtf==1.0.7",
    "configobj==5.0.9",
    "configparser==7.2.0",
    "cryptography==45.0.6",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "distro==1.9.0",
    "docx==0.2.4",
    "easygui==0.98.3",
    "ebcdic==1.1.1",
    "etelemetry==0.3.1",
    "executing==2.2.0",
    "extract-msg==0.55.0",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.7.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "groq==0.31.0",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "ipykernel==6.30.1",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.74",
    "langchain-google-genai==2.1.9",
    "langchain-groq==0.3.7",
    "langchain-openai==0.3.29",
    "langchain-pinecone==0.2.11",
    "langchain-tests==0.3.20",
    "langchain-text-splitters==0.3.9",
    "langsmith==0.4.13",
    "lark==1.1.9",
    "looseversion==1.3.0",
    "lxml==6.0.0",
    "markdown-it-py==4.0.0",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "msoffcrypto-tool==5.4.2",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nibabel==5.3.2",
    "nipype==1.10.0",
    "numpy==2.3.2",
    "olefile==0.47",
    "oletools==0.60.2",
    "openai==1.99.7",
    "orjson==3.11.1",
    "packaging==24.2",
    "pandas==2.3.1",
    "parso==0.8.4",
    "pathlib==1.0.1",
    "pcodedmp==1.2.6",
    "pillow==11.3.0",
    "pinecone==7.3.0",
    "pinecone-plugin-assistant==1.7.0",
    "pinecone-plugin-interface==0.0.7",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "prov==2.1.1",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "puremagic==1.30",
    "py-cpuinfo==9.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydot==4.0.1",
    "pygments==2.19.2",
    "pymupdf==1.26.3",
    "pyparsing==3.2.3",
    "pytest==8.4.1",
    "pytest-asyncio==0.26.0",
    "pytest-benchmark==5.1.0",
    "pytest-codspeed==4.0.0",
    "pytest-recording==0.13.4",
    "pytest-socket==0.7.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pywin32==311",
    "pyxnat==1.6.3",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "rdflib==7.1.4",
    "red-black-tree-mod==1.22",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rsa==4.9.1",
    "rtfde==0.1.2.1",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "simplejson==3.20.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "syrupy==4.9.1",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "traits==7.0.2",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3<2",
    "uvicorn>=0.35.0",
    "vcrpy==7.0.0",
    "wcwidth==0.2.13",
    "win-unicode-console==0.5",
    "wrapt==1.17.2",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
